+++ protected/extensions/yiibooster/widgets/TbButtonGroup.php
@@ -6,9 +6,12 @@
  * @copyright Copyright &copy; Christoffer Niska 2011-
  * @license [New BSD License](http://www.opensource.org/licenses/bsd-license.php) 
  * @since 0.9.10
+ * 
+ * @author Amr Bedair <amr.bedair@gmail.com>
+ * @since v4.0.0 - upgraded to bootstrap 3.1.1
  */
 
-Yii::import('bootstrap.widgets.TbButton');
+Yii::import('booster.widgets.TbButton');
 
 /**
  *## Bootstrap button group widget.
@@ -17,8 +20,8 @@ Yii::import('bootstrap.widgets.TbButton');
  *
  * @package booster.widgets.forms.buttons
  */
-class TbButtonGroup extends CWidget
-{
+class TbButtonGroup extends CWidget {
+	
 	// Toggle options.
 	const TOGGLE_CHECKBOX = 'checkbox';
 	const TOGGLE_RADIO = 'radio';
@@ -27,13 +30,18 @@ class TbButtonGroup extends CWidget
 	 * @var string the button callback type.
 	 * @see BootButton::buttonType
 	 */
-	public $buttonType = TbButton::BUTTON_LINK;
+	public $buttonType = TbButton::BUTTON_BUTTON;
 
 	/**
 	 * @var string the button type.
 	 * @see BootButton::type
 	 */
-	public $type;
+	public $context = TbButton::CTX_DEFAULT;
+	
+	/**
+	 * @var boolean indicates whether to use justified button groups
+	 */
+	public $justified = false;
 
 	/**
 	 * @var string the button size.
@@ -80,16 +88,22 @@ class TbButtonGroup extends CWidget
 	 *
 	 * Initializes the widget.
 	 */
-	public function init()
-	{
-		$classes = array('btn-group');
-
+	public function init() {
+		
+		$classes = [];
+		
 		if ($this->stacked === true) {
 			$classes[] = 'btn-group-vertical';
+		} else {
+			$classes[] = 'btn-group';
 		}
 
 		if ($this->dropup === true) {
 			$classes[] = 'dropup';
+		}
+		
+		if ($this->justified === true) {
+			$classes[] = 'btn-group-justified';
 		}
 
 		if (!empty($classes)) {
@@ -104,7 +118,7 @@ class TbButtonGroup extends CWidget
 		$validToggles = array(self::TOGGLE_CHECKBOX, self::TOGGLE_RADIO);
 
 		if (isset($this->toggle) && in_array($this->toggle, $validToggles)) {
-			$this->htmlOptions['data-toggle'] = 'buttons-' . $this->toggle;
+			$this->htmlOptions['data-toggle'] = 'buttons';
 		}
 	}
 
@@ -113,20 +127,29 @@ class TbButtonGroup extends CWidget
 	 *
 	 * Runs the widget.
 	 */
-	public function run()
-	{
+	public function run() {
+		
 		echo CHtml::openTag('div', $this->htmlOptions);
 
 		foreach ($this->buttons as $button) {
 			if (isset($button['visible']) && $button['visible'] === false) {
 				continue;
 			}
-
+			
+			$validToggles = array(self::TOGGLE_CHECKBOX, self::TOGGLE_RADIO);
+			if (isset($this->toggle) && in_array($this->toggle, $validToggles)) {
+				$this->buttonType = $this->toggle;
+			}
+			
+			$justifiedNotLink = $this->justified && $this->buttonType !== TbButton::BUTTON_LINK;
+			if($justifiedNotLink)
+				echo '<div class="btn-group">';
+			
 			$this->controller->widget(
-				'bootstrap.widgets.TbButton',
+				'booster.widgets.TbButton',
 				array(
 					'buttonType' => isset($button['buttonType']) ? $button['buttonType'] : $this->buttonType,
-					'type' => isset($button['type']) ? $button['type'] : $this->type,
+					'context' => isset($button['context']) ? $button['context'] : $this->context,
 					'size' => $this->size, // all buttons in a group cannot vary in size
 					'icon' => isset($button['icon']) ? $button['icon'] : null,
 					'label' => isset($button['label']) ? $button['label'] : null,
@@ -142,6 +165,9 @@ class TbButtonGroup extends CWidget
                     'tooltipOptions' => isset($button['tooltipOptions']) ? $button['tooltipOptions'] : array(),
 				)
 			);
+			
+			if($justifiedNotLink)
+				echo '</div>';
 		}
 		echo '</div>';
 	}