+++ protected/extensions/yiibooster/widgets/TbDataColumn.php
@@ -36,17 +36,19 @@ class TbDataColumn extends CDataColumn
 			$sort = $this->grid->dataProvider->getSort();
 			$label = isset($this->header) ? $this->header : $sort->resolveLabel($this->name);
 
-            $booster = Bootstrap::getBooster();
+            $booster = Booster::getBooster();
 
-			if ($sort->resolveAttribute($this->name) !== false){
-                            if($sort->getDirection($this->name) === CSort::SORT_ASC){
-                                $label .= ' <span class="'.($booster->fontAwesomeCss ? 'fa fa-sort-asc' : 'icon-chevron-down').'"></span>';
-                            } elseif($sort->getDirection($this->name) === CSort::SORT_DESC){
-                                $label .= ' <span class="'.($booster->fontAwesomeCss ? 'fa fa-sort-desc' : 'icon-chevron-up').'"></span>';
-                            } else {
-                                $label .= ' ';
-                            }
-                        }
+			if ($sort->resolveAttribute($this->name) !== false)
+				$label .= ' <span class="caret"></span>';
+			/* {
+				if($sort->getDirection($this->name) === CSort::SORT_ASC){
+               		$label .= ' <span class="'.($booster->fontAwesomeCss ? 'fa fa-sort-asc' : 'icon-chevron-down').'"></span>';
+                } elseif($sort->getDirection($this->name) === CSort::SORT_DESC){
+                	$label .= ' <span class="'.($booster->fontAwesomeCss ? 'fa fa-sort-desc' : 'icon-chevron-up').'"></span>';
+                } else {
+                	$label .= ' ';
+                }
+            } */
 
 			echo $sort->link($this->name, $label, array('class' => 'sort-link'));
 		} else {
@@ -102,6 +104,12 @@ class TbDataColumn extends CDataColumn
 			} else {
 				$filterInputOptions = array();
 			}
+			
+			if(!isset($filterInputOptions['class']) || empty($filterInputOptions['class']))
+				$filterInputOptions['class'] = 'form-control';
+			else
+				$filterInputOptions['class'] .= ' form-control';
+			
 			if (is_array($this->filter)) {
 				if (!isset($filterInputOptions['prompt'])) {
 					$filterInputOptions['prompt'] = '';