+++ protected/extensions/yiibooster/widgets/TbListView.php
@@ -14,8 +14,8 @@ Yii::import('zii.widgets.CListView');
  *
  * @package booster.widgets.grouping
  */
-class TbListView extends CListView
-{
+class TbListView extends CListView {
+	
 	/**
 	 * @var string the CSS class name for the pager container. Defaults to 'pagination'.
 	 */
@@ -23,9 +23,9 @@ class TbListView extends CListView
 
 	/**
 	 * @var array the configuration for the pager.
-	 * Defaults to <code>array('class'=>'ext.bootstrap.widgets.TbPager')</code>.
+	 * Defaults to <code>array('class'=>'ext.booster.widgets.TbPager')</code>.
 	 */
-	public $pager = array('class' => 'bootstrap.widgets.TbPager');
+	public $pager = array('class' => 'booster.widgets.TbPager');
 
 	/**
 	 * @var string the URL of the CSS file used by this detail view.
@@ -38,11 +38,11 @@ class TbListView extends CListView
 	 *
 	 * Initializes the widget.
 	 */
-	public function init()
-	{
+	public function init() {
+		
 		parent::init();
 
-        $booster = Bootstrap::getBooster();
+        $booster = Booster::getBooster();
 		$popover = $booster->popoverSelector;
 		$tooltip = $booster->tooltipSelector;