+++ protected/extensions/yiibooster/widgets/TbMenu.php
@@ -7,7 +7,7 @@
  * @license [New BSD License](http://www.opensource.org/licenses/bsd-license.php) 
  */
 
-Yii::import('bootstrap.widgets.TbBaseMenu');
+Yii::import('booster.widgets.TbBaseMenu');
 
 /**
  * Bootstrap menu.
@@ -16,12 +16,13 @@ Yii::import('bootstrap.widgets.TbBaseMenu');
  *
  * @package booster.widgets.navigation
  */
-class TbMenu extends TbBaseMenu
-{
+class TbMenu extends TbBaseMenu {
+	
 	// Menu types.
 	const TYPE_TABS = 'tabs';
 	const TYPE_PILLS = 'pills';
 	const TYPE_LIST = 'list';
+	const TYPE_NAVBAR = 'navbar';
 
 	/**
 	 * @var string the menu type.
@@ -39,6 +40,12 @@ class TbMenu extends TbBaseMenu
 	 * @var boolean indicates whether the menu should appear vertically stacked.
 	 */
 	public $stacked = false;
+	
+	/**
+	 * @var boolean indicates whether the menu should be justified.
+	 */
+	public $justified = false;
+	
 
 	/**
 	 * @var boolean indicates whether dropdowns should be dropups instead.
@@ -50,20 +57,24 @@ class TbMenu extends TbBaseMenu
 	 *
 	 * Initializes the widget.
 	 */
-	public function init()
-	{
+	public function init() {
+		
 		parent::init();
 
 		$classes = array('nav');
 
-		$validTypes = array(self::TYPE_TABS, self::TYPE_PILLS, self::TYPE_LIST);
+		$validTypes = array(self::TYPE_TABS, self::TYPE_PILLS, self::TYPE_LIST, self::TYPE_NAVBAR);
 
 		if (isset($this->type) && in_array($this->type, $validTypes)) {
-			$classes[] = 'nav-' . $this->type;
+			$classes[] = $this->type === self::TYPE_NAVBAR ? 'navbar-nav' : 'nav-' . $this->type;
 		}
 
 		if ($this->stacked && $this->type !== self::TYPE_LIST) {
 			$classes[] = 'nav-stacked';
+		}
+		
+		if ($this->justified && $this->type !== self::TYPE_LIST) {
+			$classes[] = 'nav-justified';
 		}
 
 		if ($this->dropup === true) {
@@ -72,7 +83,7 @@ class TbMenu extends TbBaseMenu
 
 		if (isset($this->scrollspy)) {
 			$scrollspy = is_string($this->scrollspy) ? array('target' => $this->scrollspy) : $this->scrollspy;
-			$this->widget('bootstrap.widgets.TbScrollSpy', $scrollspy);
+			$this->widget('booster.widgets.TbScrollSpy', $scrollspy);
 		}
 
 		if (!empty($classes)) {
@@ -92,9 +103,9 @@ class TbMenu extends TbBaseMenu
 	 *
 	 * @return string the class name
 	 */
-	public function getDividerCssClass()
-	{
-		return (isset($this->type) && $this->type === self::TYPE_LIST) ? 'divider' : 'divider-vertical';
+	public function getDividerCssClass() {
+		
+		return (isset($this->type) && $this->type === self::TYPE_LIST) ? 'nav-divider' : 'divider-vertical';
 	}
 
 	/**
@@ -104,8 +115,8 @@ class TbMenu extends TbBaseMenu
 	 *
 	 * @return string the class name
 	 */
-	public function getDropdownCssClass()
-	{
+	public function getDropdownCssClass() {
+		
 		return 'dropdown';
 	}
 
@@ -116,8 +127,8 @@ class TbMenu extends TbBaseMenu
 	 *
 	 * @return boolean the result
 	 */
-	public function isVertical()
-	{
+	public function isVertical() {
+		
 		return isset($this->type) && $this->type === self::TYPE_LIST;
 	}
 }