+++ protected/extensions/yiibooster/widgets/TbProgress.php
@@ -6,7 +6,7 @@
  * @copyright Copyright &copy; Christoffer Niska 2011-
  * @license [New BSD License](http://www.opensource.org/licenses/bsd-license.php)
  */
-
+Yii::import('booster.widgets.TbWidget');
 /**
  *## Bootstrap progress bar widget.
  *
@@ -15,17 +15,8 @@
  * @package booster.widgets.decoration
  * @since 0.9.10
  */
-class TbProgress extends CWidget
-{
-	const TYPE_INFO = 'info';
-	const TYPE_SUCCESS = 'success';
-	const TYPE_WARNING = 'warning';
-	const TYPE_DANGER = 'danger';
-
-	/**
-	 * @var string the bar type. Valid values are 'info', 'success', and 'danger'.
-	 */
-	public $type;
+class TbProgress extends TbWidget {
+	
 	/**
 	 * @var boolean indicates whether the bar is striped.
 	 */
@@ -55,48 +46,44 @@ class TbProgress extends CWidget
 	 * @var array $stacked set to an array of progress bar values to display stacked progress bars
 	 * <pre>
 	 *  'stacked'=>array(
-	 *      array('type' => 'info|success|warning|danger', 'percent'=>'30', 'htmlOptions'=>array('class'=>'custom')),
-	 *      array('type' => 'info|success|warning|danger', 'percent'=>'30'),
+	 *      array('context' => 'info|success|warning|danger', 'percent'=>'30', 'htmlOptions'=>array('class'=>'custom')),
+	 *      array('context' => 'info|success|warning|danger', 'percent'=>'30'),
 	 *  )
 	 * </pre>
 	 * @since 9/21/12 8:14 PM antonio ramirez <antonio@clevertech.biz>
 	 */
 	public $stacked;
 
+	protected $progressClasses = array('progress');
+	protected $progressBarClasses = array('progress-bar');
+	
 	/**
 	 *### .init()
 	 *
 	 * Initializes the widget.
 	 */
-	public function init()
-	{
-		$classes = array('progress');
-		if (empty($this->stacked)) {
-			$validTypes = array(self::TYPE_INFO, self::TYPE_SUCCESS, self::TYPE_WARNING, self::TYPE_DANGER);
+	public function init() {
+		
+		if ($this->isValidContext())
+			$this->progressBarClasses[] = 'progress-bar-' . $this->getContextClass();
+		
+		if ($this->striped)
+			$this->progressClasses[] = 'progress-striped';
+		
+		if ($this->animated)
+			$this->progressClasses[] = 'active';
 
-			if (isset($this->type) && in_array($this->type, $validTypes)) {
-				$classes[] = 'progress-' . $this->type;
-			}
-			if ($this->striped) {
-				$classes[] = 'progress-striped';
-			}
-			if ($this->animated) {
-				$classes[] = 'active';
-			}
+		if ($this->percent < 0)
+			$this->percent = 0;
+		else if ($this->percent > 100)
+			$this->percent = 100;
 
-			if ($this->percent < 0) {
-				$this->percent = 0;
-			} else if ($this->percent > 100) {
-				$this->percent = 100;
-			}
-		}
-
-		if (!empty($classes)) {
-			$classes = implode(' ', $classes);
+		if (!empty($this->progressClasses)) {
+			$this->progressClasses = implode(' ', $this->progressClasses);
 			if (isset($this->htmlOptions['class'])) {
-				$this->htmlOptions['class'] .= ' ' . $classes;
+				$this->htmlOptions['class'] .= ' ' . $this->progressClasses;
 			} else {
-				$this->htmlOptions['class'] = $classes;
+				$this->htmlOptions['class'] = $this->progressClasses;
 			}
 		}
 	}
@@ -108,11 +95,11 @@ class TbProgress extends CWidget
 	 * @since  9/21/12 8:13 PM  antonio ramirez <antonio@clevertech.biz>
 	 * Updated to use stacked progress bars
 	 */
-	public function run()
-	{
+	public function run() {
+		
 		echo CHtml::openTag('div', $this->htmlOptions);
 		if (empty($this->stacked)) {
-			echo '<div class="bar" style="width: ' . $this->percent . '%;">' . $this->content . '</div>';
+			echo '<div class="'.implode(' ', $this->progressBarClasses).'" style="width: ' . $this->percent . '%;">' . $this->content . '</div>';
 		} elseif (is_array($this->stacked)) {
 			foreach ($this->stacked as $bar) {
 				$options = isset($bar['htmlOptions']) ? $bar['htmlOptions'] : array();
@@ -128,11 +115,20 @@ class TbProgress extends CWidget
 				} else {
 					$options['style'] .= ' ';
 				}
-				$options['class'] .= 'bar bar-' . $bar['type'];
+				$options['class'] .= 'progress-bar progress-bar-' . $bar['context'];
 
 				echo '<div ' . CHtml::renderAttributes($options) . '>' . @$bar['content'] . '</div>';
 			}
 		}
 		echo CHtml::closeTag('div');
+	}
+	
+	protected function isValidContext($context = false) {
+		return in_array($this->context, [
+			self::CTX_SUCCESS, 
+			self::CTX_INFO, 
+			self::CTX_WARNING, 
+			self::CTX_DANGER
+		]);
 	}
 }